+++ legadilo/feeds/models/__init__.py
@@ -5,7 +5,6 @@
 from .feed import Feed
 from .feed_article import FeedArticle
 from .feed_category import FeedCategory
-from .feed_deleted_article import FeedDeletedArticle
 from .feed_tag import FeedTag
 from .feed_update import FeedUpdate
 
@@ -13,7 +12,6 @@ __all__ = [
     "Feed",
     "FeedArticle",
     "FeedCategory",
-    "FeedDeletedArticle",
     "FeedTag",
     "FeedUpdate",
 ]